Move test const file
